Scrollable confirmation dialogs
Add a conditional scrollbar to confirmation dialogs when
the dialog is large. This prevents a dialog from having an
extreme vertical height that can stretch off screen.

The condition for scrollbars is not extremely precies, uses
the total message length in characters as a measure for
how large the contents are, anything over 1000 characters
will get scrollbars.