Rename 'DownloadedMaps' -> 'installedMaps'
Also renames 'DownloadedMapsListing' -> 'InstalledMapsListing'

The reason for this rename is to make it more clear that these
are installed maps that have already been downloaded and are
not recent downloads or merely available to download. The difference
in terms makes it easier to distinguish from maps that are available
for download.