Put map download listing behind an interface 'MapsListingClienting'
Puts the fetch of map download listing from YAML file and the
fetch for download listing from maps server behind the same interface.
This simplifies how we fetch map download listings and now we only
need to select the right implementation based on feature flag
that tells us whether to go to maps (lobby) server or whether
to download and parse YAML file.