Fix delete of maps, use and add a FileUtiles#deleteDirectory method
Deleting non-empty directories fails, now that maps are directories and
no longer zip files, the delete command that was used to delete the zip
fail can fail when used to delete the map directory.

This update solves the problem by using a library to delete the directory
which does a recursive delete first of all files and directories first.