Consolidate 'run_ansible' scripts
Merge run_nasible scripts together and use script args
to control which environment we deploy to.

As part of this the 'provision' one-time setup step
is merged into the main deployment. There is now a
'full' and 'minimal' deployment, where the minimal
is done post-merges and the full is reserved
to be run by hand. 'full' is appropriate if
there are new servers to be configured.