Battle Calculator Make shortcut CTRL-B work
So far the shortcut CTRL-B to open a new battle calculator only works in the
main TripleA window.  Add this shortcut also to the battle calculator dialog.

Note: The new battle calculator is opened for the currentTerritory which is the
territory the mouse was over last in the main TripleA window.  Mouse movements
in dialogs are not seen by the main window and thus do not update the current
territory.