Battle Calculator: Focus calculate odds button
The calculate odds button is not focused because it is disabled
at the time requestFocus() is called.  Move that call to the site
where the button is enabled, so requesting the focus can succeed.

As a side effect, this fixes the focusing of th battle calculator.
Prior to this change, the calculator immediatly lost focus once it
was created, now it keeps the focus.

Also instead of using the discouraged requestFocus() use
requestFocusInWindow().