Move new Battle Calculator dialogs always to the front
On some desktop environments, there is an odd behaviour: The new battle
calculator dialog is not put at the front if the last user-interaction was to
move the window of an already existing battle calculator dialog.  Oddly enough,
calling toFront() directly before or after setVisible(true) has no effect
either. However, delaying the call to the end of the queue of the Event
Dispatch Thread using SwingUtilities.invokeLater() solves the issue (though you
can see the new dialog in the background for an blink of an eye).

Tested with Cinnamon Desktop 4.8.5.