Battle Calculator: Return NaN if no results were aggregated
Let the methods of AggregateResults return NaN if no results were
added to that instance. Previously, 0 was returned which is equally
arbitrary.

Note that all call sides from the AI ensure that results are added
or they don't call methods like ...IfAttackerWon() where its possible
that no results remain to calculate the mean on.