Fix map update, replace existing maps on update
Map update was written to not expect an unzipped map to already
exist. In the case of map update, then there will be an already
existing map folder. This update fixes the update to do a replace
of any existing content.

The folder replace is done with a backup and rollback if it fails.
Before we replace the old map, we will move it to temp. If the
move of the new map fails, then we will move the old map
from the temp backup to its original location.