Dev-Config: Disable code climate too many return statements warning. (#9625)
Turning this check off, we'll know too many return statements
code smell when we see it. In principle we want a check that
verifies the code paths to a methods return statements
are easy to understand, counting the number of return statements
alone is a somewhat poor proxy for that.