OpponentSelector: Prioritise the territory owner
If the territory owner did not have any units in its territory, it was treated
as an unrelated other player.  Make the owner a priority player with the lowest
priority of all players.  This way the owner is still prefered for an empty
territory but ignored if units are in the territory.