Update: Move map.yaml generation to map extraction
Instead of generating a missing 'map.yaml' file when simply
reading maps in the maps folder, we will geneate missing 'map.yaml'
files if we are extracting to a maps folder.

The extraction process is meant to help support upgrading from 2.5,
it is this time when we really expect for 'map.yaml' files to be missing.