Fix 'currentPlayer' variable may be null warning
In the private method 'currentPlayer' can be be null and is passed to a 'List.of' which
would cause a NPE. Though, we only call the method from one place and at that time
we guarantee that currentPlayer is non-null.

Two problems to solve:
A. It was not obvious that 'currentPlayer' was actually guaranteed to be non-null
B. If the naive coder were to relay a particular get method and invoke this method
without guaranteeing the precondition that 'currentPlayer' is not null, then there
would be a NPE.

To resolve these two problems we inline the method. It makes it clear from the context
where the method is invoked that 'currentPlayer' is not null, and there is no liability
to invoke a private method with improper state (removes that as a possibility for error)