Toggle 'use cache' = false for ImageIO -> triggers to use in-memory cache
'use cache' refers to whether to use a disk cache or in-memory.

When set to false we use more memory and less disk-IO, when set to
true the inverse. Overall memory usage did not seem to differ that
much between the two.

Runtime performance for loading images with 'cache = false' is
better with typical (and very consistent) load times of 0.1ms.
When set to false the typical load time was more like 0.4ms and
had a lot more extreme values above 1.0ms.