Fix (remove) hanging game data lock release
Lock acquisition was removed in a prior commit, if we remove a lock
we have not acquired there is an error. This update fixes this by
removing the lock release (rather than adding back in the lock
acquisition).

Looking over where we are locking, we are accessing immutable data
and do not need to lock game data.

Reported in, and fixes: #9890