Update failure to connect to lobby as a reportable-error
Rename 'showError' to 'showMessage' in lobby and use 'logger' to convey unable
to connect to lobby.

This will be of mixed benefit because we de-dupe errors once per version of the
game. Nonetheless, even if users are commenting on a closed issue, that can still
be helpful to notify us of problems.