Header redirect/lobby variable port number (#9940)
* Nginx Config: remove deprecated ssl property

'sudo nginx -t' was generating this warning:

> nginx: [warn] the "ssl" directive is deprecated, use the "listen ... ssl" directive instead in /etc/nginx/sites-enabled/default:21

* Nginx config to redirect to differing lobby depending on header version

Adds an NGINX config that checks the 'triplea-version' header value,
and based on that value redirects. To support running multiple servers
on the same host, we need to parameterize the lobby server to be able
to inject port number. The end goal is we can have multiple lobbies
running on the same host, different port numbers, and clients will
be redirected to the correct lobby based on version number in header.