Remove mitogen, a plugin for ansible that increases deployment speed
Mitogen delivers dramatically increased deployment speed for ansible
deployments. Despite this, the installation step is failing due
to an expired https certificate while attempting to download the
plugin.

We are also missing the 'ansible.cfg' configuration that uses mitogen,
so we are not getting the benefit of mitogen. Mitogen also has a drawback
that it pins us to an older ansible version.

Given these trade-offs, this update removes mitogen.