Disk usage: stop automatically downloading maps on bots (#10248)
The download script for maps on bots is a bit odd now as it does not
take into account that maps are unzipped. This results in repeated downloads
of map zips which then remain on the system occupying significant disk
space. This update removes this functionality until it can be altered in some
way to avoid disk space issues.