Remove unused Kotlin dependency & plugin
We do not yet have any dependencies on Kotlin that make use of
the checked in Kotlin build dependency. The Kotlin compilation
step is observed to execute on all builds and takes a few seconds
(relatively long time for a process that overall takes a couple dozen
seconds). This update removes the unused build plugin and
dependencies.