Fix selecting step nodes in history view not having an effect.
Fix selecting step nodes in history view not having an effect.

This was broken with https://github.com/triplea-game/triplea/pull/10429, which made the logically correct change, but ran into a bug where we weren't actually setting the change index on step nodes. Introduce a fix as well as handling of history from saved games that didn't set it.