Fix errors with map resources being not found in 2.6.
This was caused by some maps having duplicate files (e.g. a polygon.txt file in baseTiles folder). Before this change, it's undefined which of these files will be found first and all other files will be expected to be alongside it.

With this change, the search returns files in a deterministic order, sorted by path length. So files closer to the root will be first in the list. This makes the behavior deterministic across platforms and the map to use the file closest to the map root.