Fix unit chooser missing "All" button.
A 2.6 change resulted in a Predicate being passed to the UnitChooser, which caused the All button to be hidden. That was not intentional and this PR makes the visibility of that button get set explicitly, converting call sites that want that button hidden to specify so.

Additionally, code for the None button, which was never shown, is cleaned up.