Vagrant: Conditionally generate dhparams file
It takes a significant amount of time to generate dhparams file (around 10-20 minutes).
It is a one-time activity, but on vagrant we wind up doing this often because we
frequently recreate the VM. This update makes the file generation to be conditional
that we are not in a vagrant environment, if deploying to vagrant then we deploy
a pregenerated file to save time.