Refactor TuvUtils to allow caching to speed up AIs & battle calculator. (#10680)
Refactor TuvUtils to allow caching. New TuvCostsCalculator class is introduced, that keeps around intermediate calculations allowing repeated calls to re-use them. Refactors call sites to call the new class, although the BattleCalculator call site is the one that benefits the most from this, by re-using the same object for all battle simulations (which is heavily used by AI).

This should be especially impactful for maps with lots of units, for example "another world". I tested a round 7 game on "another world" with all Hard AI players and the mean runtime of MustFightBattle.fight() went from 3.0ms to 2.3ms, a 23% speedup (which was called around ~185,000 times during the course of AI simulations).