Simplify servers.yml pending different version handling
Servers.yml will eventually no longer be used in favor of the server handling
correct version routing. Further, we are likely to have one server moving
forward that will enforce its own 'min-version'.

This update simplifies the 'servers.yml' file. Eventually we could delete it,
though the path to the file is hardcoded in 2.5 and previous game-clients