Code-Climate Config: exclude markdown (md) files
It does not make a lot of sense to do markdown format linting, so it is
being turned off. We routinely ignore the called out warnings and they
are largely not important. The MD syntax enforces a consistent enough style,
beyond this deviations in consistency of MD files is largely inconsequential.
So, this is just not an important thing for us to be linting on right now.
Perhaps in the future an auto-formatter could do all the formatting for us.