Dev-UX: Update verify script to check for docker dependency
We modify the verify script to look for 'docker' so that we can
fail early with a more helpful error message. Future improvements
might try to auto-install docker or at least provide a link to
install in the error output.

Now that the verify script is more than just a couple of commands,
We also change the structure of the 'verify' script to have a 'main'
block.