Simplify & Fix: Properly skip database tests before database has been started
The verify script has a command to start database, but this is done only after
we have run 'check' in parallel on all sub-modules except for those that
depend on tests. The database dependent tests were specified down to the exact
sub-module, which did not scale as we added a sub-module but did not update
this white list.

The result of having a test start before database is that we got an error message
that the database was not available. This gave the false impression that we
need to start database before running 'verify' (when really it was just broken).

This update resolves this by specifing the lobby top level module that we will
skip without specifying each individual sub-module below the 'lobby top level'.

To better root out any problems with database starting up, we move that to be
started first before we execute any checks.