Simplify passing API key & Headers to Feign Clients (#10714)
* Move class 'AuthenticationHeaders' to 'Lobby-Client'

The class 'AuthenticationHeaders' is used to create headers that are
specific to game-clients communicating with the lobby. Previously
'AuthenticationHeaders' was in the very generic 'http-client' package,
which is not appropriate since the class is server specific.

* Specify headers when creating HTTP client instead of when invoking each endpoint

* Simplify, combine 'feign http clients' into their wrapper class

The wrappers around the 'feign http clients' were there mostly
to hide passing of headers to endpoints. Now that we have removed
the need to pass headers to each individual endpoint, the wrappers
provide nothing extra. This update simplifies by essentially inlining
the wrapped http interfaces into their wrappers.