Ansible: Reconfig for more modular lobby_server configuration
Make the lobby_server role more module. Do this primarily
by including other roles and passing variables to the
included roles.

We also drastically simplify the way that nginx certs are loaded.
We now keep variables for where we expect to have certificates
and we override them for vagrant & prod. We then pair this
with another variable that tells us whether we should
auto-generate certificates, or run certbot; and we similarly
set these values in prod & vagrant override configs.

Ansible role break-out:
  database/postgres -> { database/postgres/install, database/postgres/create_datebase }
  database/flyway -> { database/flyway/install, database/flyway/run_migrations }
  nginx -> { nginx/install, nginx/generate_dhparam, nginx/generate_self_signed_cert }
  certbot -> { nginx/run_certbot }

By using better flagging for when we execute certbot, we no longer
need the 'letsencrypt' hostgroup and the certbot role is now
executed as appropriate along with the lobby-deploy role