Drop Codacy & Rename and combine github workflow actions
(1) Drops codacy - the static analysis is useful but in practice we tend
to ignore the codacy warnings. Usually PMD will pick up any problems that
codacy highlights, otherwise it is usually problems in existing code
that are flagged.

(2) Combine workflow actions together and organize them.
Static code checks are prefixed with 'check-*', testing tasks are prefixed with
'testing-*', and all on-merge-to-master release actions are now in 'do-release.yml'

Actions that are run on-demand, are now prefixed with 'run-*'