Disconnect github action to update maps on website
We have some configuration to automatically split the 'triplea_maps.yaml'
file (which indexes avaialable maps) and publishes these map descriptions
to the TripleA website.

This has a few problems:
1. We are quickly moving away from 'triplea_maps.yaml' and will need
a different mechanism if we wish to continue updating maps on website.

2. The website receives exceedingly low traffic, it's likely nobody
has even looked at the maps listing on the website for a few weeks.
The value to complexity ratio is seemingly unfavorable for even building
a replacement, which is all more argument for removing the current legacy
implementation.