Fix 'run_ansible': define variable and print ansible-playbook command
A variable was needed for the vault password file location and was
erronously removed previously. This update restores the variable.

We also fix some relative paths in 'run_ansible', instead of assuming
a current directory, we explicitly use a '$rootDir'

Last, use 'set -x' to print the ansible-playbook command.
Given that 'run_ansible' is just a wrapper around the ansible-playbook
command, it is particularly useful to see the ansible-playbook command
in its full as its being run.