Github Acitons: Use webfactory/ssh-agent to Load Private SSH Key into SSH-Agent
We're seeing a failure to add our SSH key to ssh-agent (ssh-agent probably not running).
"Could not open a connection to your authentication agent."

This update adds an action to do this for us, loading the SSH key from a secrets variable:
https://github.com/webfactory/ssh-agent

Previously we used a vault password to decrypt a checked in private key and then load
that into ssh-agent. Now that we will obtain the private key from the config of github
action (added via repository settings), we no longer need the encrypted private key
that was checked in.