Simplify: update ClientLoginValidator to be constructed with a builder
We can fold in 'serverMessenger' as a constructor arg with password
being optional.

We also avoid passing in the 'engine' version and instead read it
directly. This avoids mixed usage where sometimes we read the engine
version and sometimes we used the passed in value. Not passing
in the engine version value simplifies construction.