Infra/sha in version (#10796)
* Remove check on save games for compatibility

- We will just try to load the save game
- If there are any problems, a simple warning will be printed with
error message.

Previous behavior aborted the load entirely. There was a 'default = false'
option put in to load save games anyways, this update effectively
makes that option always true.

* Simplify, remove wiring of engine version parameter

'engine version' variable is no longer used when cloning
a game data, this leads to a cascade of simplifications
where we do not need to pass an engine version parameter.

When writing data, we can use the 'ProductVersionReader'
directly avoid the additional wiring.

* Version Number: split concept of build & product version, add SHA to Version NUmber

This update primarily:
- splits concept of the version number, there is now 'product version', and a 'build version'
- build version is no longer suffixed with the CI 'run number', instead it is suffixed
  with git SHA.
- creates an nginx mapping of product version to lobby port number

Version numbers will now look something like: "2.6+ea007xyz"

The split of 'product version' and 'build version' is for more for clarity, the two version
values are different but related. This split is helpful as the lobby deployment needs to
know about the 'build version' so it knows the name of the 'jar' files it will deploy.
Though, it also now needs to know about the product version so it knows which port
number to use. Splitting the two concepts keeps this easier and avoids confusion about
reading a pre, or post-modified version of 'product.properties'.

To further make this clear, we want an idempotent way to get the build or product version.
Before, we did not do this in an idempotent way and it lead to the above confusion regarding
'product.properties', and extra effort to get the product version back from a build version.

To resolve this, we move the 'source of truth' version number into the build, '.build/product-version.txt'
now contains the build number, and we use the deterministic 'git SHA" to build the release number.
This is further useful, as before we had a tag that matched the version so that we can find the
same code version easily. By having the git sha, we can still find the exact same code version.