Drop code-climate (#10780)
code-climate is a static analysis tool that automatically adds comments in PRs
to flag static check violations. We can customize things like line counts,
arg counts etc.. that constitute violations.

While such checks are really useful, on the whole it is causing more noise
than is helping. Most updates are to existing code that typically has several
violations in almost any module that is being updated. So, we tend to ignore
any flagged violations, or we are improving the violations but still get flagged
due to being above threshold.

If there were a way to have code climate run for only exactly modified code,
or if we had a new code base, then it would be much more useful. Otherwise
it is important to avoid PR checks that can be ignored, a 'red build' should
generally mean 'not suitable for merge'