Add prefs for two combat warnings.
The following two similar warnings are given their own prefs:
  1. Warning about flying over an enemy territory that has AA
  2. Warning about attacking a territory to which planes can scramble

A "Don't show this warning" checkbox is added to each of the dialogs, as well as a corresponding pref setting (one for each).
Additionally, the second warning is now controlled by this preference setting, rather than the "beta features" setting.