Fix NPE in ObjectivesPanel
NPE happens trying to launch any game. Problem is that 'getName' is called before
the constructor in this class, in this case 'resourceLoader' is not yet initiatilized
and we then go on to have a NPE in 'ObjectiveProperties'

The override of 'getName' for a swing component is generally not necessary, this override
can be removed entirely and it fixes this problem.