Remove PropertyFile.java (removes cache of 'new Properties(..)') (#10841)
* Simplify, remove convenience constructor in PropertyFile

* Simplify: remove class 'PropertyFile'

PropertyFile largely provides a cache around constructing a 'Properties' object.
Removing this simplifies. The benefit of the cache was slight in absolute
numbers, reducing a 100 micro-second operation to 10. The call volume is not
so large either that this adds up and provides a big savings. The expensive
part of 'Properties' still is accessing  individual property values, which is
already a much larger time than constructing the Properties. So this cache
does not make much sense.