Infra: ensure artifacts are built when deploying (#10844)
This updates the build artifact script to ensure we always build
the expected artifacts. Gradle will not build artifacts if they
appear up-to-date. If the git-SHA then changes, the expected
artifact name could be different from existing artifacts (that
potentially were previously built but are not rebuilt because
they appear up to date).

So, this update checks if the expected artifact names exist,
if not, we do a clean to remove any previous artifacts and
then build fresh artifacts that then should have the expected
names.