Infra: use commit number as build number
Instead of using git commit sha, this update uses the commit number
instead. So we would have version numbers like "2.6+16243" instead
of "2.6+aed412"

A benefit of a commit number is it is deterministic like commit
sha, but also sequential. We still will be able to find a specific
code version by 'version number' since each build is tagged
and we can checkout the tag (so we do not lose a capability compared
to using git SHA, a version still uniquely identifies a build)