Simplify build artifacts script
We no longer need to do an explicit clean to remove prevoiusly built artifact.
We had that in place because the previous artifact might have looked the same
to gradle (and hence not be built, because not changed), but could have had
a different name. Now that version number is no longer in the name, we are
guaranteed to find artifacts built by gradle.