Infra: run all commands with 'become' (root) (#10870)
An update to ansible.cfg such that ansible will automatically
add a "become: true" to all commands. This means we run all
commands with "sudo" (effectively root). In practice this has
no effect because right now we connect to all servers as root
and run ansible commands as root user anyways.

This update is helpful if we ever change our connection user,
if we ever connect and run ansible commands as a user that is not
root.

When running deployments from local, we can use the 'admin'
user and our pre-installed keys to successfully run deployments.
In such a case our SSH user does change and having the 'become: true'
is useful. Notably, some commands require 'sudo' and we are able
to run them now only because we are connected as root.