New Map - Modern Warfare
Hey, I wish to share my new map, Modern Warfare, with the community. I have tried transfering ownership in order to make the repository show up in "triplea - maps", but it didn't work. I hope it is ok to do it this way. If it doesn't, let me know what I can do to fix it. Thanks :)