Fix merge resolution, remove 'version' from build.gradle
The property cascades down to all shadow jar commands and winds
up with build artifacts (jar files) that have version number in
them. Instead, we only want the 'game-headed' artifact to have
a version number in it. The other artifacts are server side and
are much easier to deal with if the jar files have names that
do not change.