Infra: fix version number on game-headed build artifacts
Version number on build artifacts (install4j installers) had
their version number printing as 'unspecified'. This update
fixes that problem. In previous commits we remove the 'version'
variable from gradle in favor of specifying it more precisely
in just game-headed (this allowed for other projects to build
jar files that do not include a version number in them). It
was omitted to update the install4j task to account for this.